refactored functions and services
