--- src/utils/classes/Fullhouse.js
@@ -1,3 +1,0 @@
-class Fullhouse {}
-
-export default Fullhouse
