--- src/utils/classes/Result.js
@@ -1,3 +1,0 @@
-class Result {}
-
-export default Result
